Rename a namespace
