+++ spec/Knp/Rad/User/EventListener/SaltGenerationListenerSpec.php
@@ -4,7 +4,7 @@ namespace spec\Knp\Rad\User\EventListener;
 
 use Doctrine\ORM\Event\LifecycleEventArgs;
 use Knp\Rad\User\Salt\Generator;
-use Knp\Rad\User\User\HasSalt;
+use Knp\Rad\User\HasSalt;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
