+++ src/Knp/Rad/User/EventListener/PasswordGenerationListener.php
@@ -5,7 +5,7 @@ namespace Knp\Rad\User\EventListener;
 use Doctrine\ORM\Event\LifecycleEventArgs;
 use Knp\Rad\User\Password\Generator;
 use Knp\Rad\User\Password\Generator\HackzillaGenerator;
-use Knp\Rad\User\User\HasInitialPassword;
+use Knp\Rad\User\HasInitialPassword;
 
 class PasswordGenerationListener
 {
